Revert "Separate header and source files"
This reverts commit 7ffd6c02623aa2d46203e75492962de74c560f60.
