+++ CMakeLists.txt
@@ -10,17 +10,8 @@ FetchContent_Declare(SFML
     GIT_TAG 2.6.x)
 FetchContent_MakeAvailable(SFML)
 
-add_executable(CMakeSFMLProject
-    src/hudhair.cpp
-    src/josh_b.cpp
-    src/luc.cpp
-    src/main.cpp
-    src/player.cpp
-    src/tyler.cpp
-    src/scene/Component.cpp
-    src/scene/Entity.cpp
-)
-target_link_libraries(CMakeSFMLProject PRIVATE sfml-graphics sfml-window sfml-system)
+add_executable(CMakeSFMLProject src/main.cpp)
+target_link_libraries(CMakeSFMLProject PRIVATE sfml-graphics)
 target_compile_features(CMakeSFMLProject PRIVATE cxx_std_17)
 
 if(WIN32)
