+++ src/josh_b.cpp
@@ -1,20 +1,28 @@
-#include "josh_b.hpp"
+#include <SFML/Graphics.hpp>
+#include <SFML/Graphics/CircleShape.hpp>
 
-JoshB::JoshB(sf::RenderWindow& window) : window(window), circle(200) {}
+class JoshB {
+public:
+    JoshB(sf::RenderWindow& window) : window(window), circle(200) {}
 
-void JoshB::draw() {
-    // change the number of sides (points) to 100
-    circle.setPointCount(125);
-    window.draw(circle);
-}
+    void draw() {
+        // change the number of sides (points) to 100
+        circle.setPointCount(125);
+        window.draw(circle);
+    }
 
-void JoshB::update() {
-    // set the absolute position of the entity
-    // circle.setPosition(10.f, 50.f);
+    void update() {
+        // set the absolute position of the entity
+        // circle.setPosition(10.f, 50.f);
 
-    // move the entity relatively to its current position
-    circle.move(0.02f, 0.05f);
+        // move the entity relatively to its current position
+        circle.move(0.02f, 0.05f);
 
-    // retrieve the absolute position of the entity
-    // sf::Vector2f position = circle.getPosition(); // = (15, 55)
-}
+        // retrieve the absolute position of the entity
+        // sf::Vector2f position = circle.getPosition(); // = (15, 55)
+    }
+
+private:
+    sf::RenderWindow& window;
+    sf::CircleShape circle;
+};
