+++ src/luc.cpp
@@ -1,20 +1,28 @@
-#include "luc.hpp"
+#include <SFML/Graphics.hpp>
+#include <SFML/Graphics/CircleShape.hpp>
 
-Luc::Luc(sf::RenderWindow& window) : window(window), circle(200) {}
+class Luc {
+public:
+    Luc(sf::RenderWindow& window) : window(window), circle(200) {}
+    
+    void draw() {
+        // change the number of sides (points) to 100
+        circle.setPointCount(75);
+        window.draw(circle);
+    }
 
-void Luc:: draw() {
-    // change the number of sides (points) to 100
-    circle.setPointCount(75);
-    window.draw(circle);
-}
+    void update() {
+        // set the absolute position of the entity
+        // circle.setPosition(10.f, 50.f);
 
-void Luc::update() {
-    // set the absolute position of the entity
-    // circle.setPosition(10.f, 50.f);
+        // move the entity relatively to its current position
+        circle.move(0.10, 0.05f);
 
-    // move the entity relatively to its current position
-    circle.move(0.10, 0.05f);
+        // retrieve the absolute position of the entity
+        // sf::Vector2f position = circle.getPosition(); // = (15, 55)
+    }
 
-    // retrieve the absolute position of the entity
-    // sf::Vector2f position = circle.getPosition(); // = (15, 55)
-}
+private:
+    sf::RenderWindow& window;
+    sf::CircleShape circle;
+};
